[monaco-textmate] Added new Syntax Coloring
Converted the weird singleton ThemeService to an actual singleton instance managed by inversify.
Good thing is: Now it can be injected with other component, and is better integrated with the rest.
Bad thing is: It might not be as fast as before on startup, but the lag is negligeable
   (maybe this will be the opportunity to design a faster pre-init sequence)

ThemeService now also depends on `ThemeProvider`s to provide themes.
It is asynchronous so that we can load themes from disk in the future.

The monaco extension also relies on `MonacoThemeProvider`s now to fetch themes for the `monaco-editor`.
Because themes work in three steps:
 1. We define a theme for Theia's UI, and we provide the editor theme as a string id.
 2. We apply the editor theme based on the string id.
 3. We trigger the `activate` and `deactivate` callbacks.
The changes made in monaco are meant to provide the themes assiociated with the id.

New extension: `monaco-textmate`, that provides the logic requires for Textmate to work in monaco.
It also specifies new themes, both for Theia and for the `monaco-editor`.
`monaco-textmate` also wires the Textmate token provider for monaco and the different grammars to parse the source files.

Signed-off-by: Patrick-Jeffrey Pollo Guilbert <patrick.pollo.guilbert@ericsson.com>
Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>