!fixup Faster Theme loading at startup
Loading the theme as fast as possible is not always possible.
In a perfect world, we would like a Theme to be applied instantly.

In reality, because themes can be contributed by multiple extensions, we may have to wait a bit.

The best solution would be to have a generic startup theme as a placeholder, while the actual themes are being loaded.

This commit is a step in this direction.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>